+++ examples/polynomial2D.py
@@ -10,8 +10,8 @@ from sao.intervening_variables import MMAp, MixedIntervening
 from sao.intervening_variables.mma import MMA02 as MMA
 from sao.move_limits import Bounds, MoveLimit
 from sao.problems import Subproblem
-from sao.solvers.SolverIP_Svanberg import ipsolver
-from sao.solvers.cvxopt_wrapper import cvxopt_solver
+from sao.solvers.pdip_svanberg import ipsolver
+from sao.solvers.wrappers.cvxopt import cvxopt_solver
 from sao.util import Plot
 
 # Set options for logging data: https://www.youtube.com/watch?v=jxmzY9soFXg&ab_channel=CoreySchafer
