+++ sao/approximations/__init__.py
@@ -1,2 +1,4 @@
 from .approximation import Approximation
 from .taylor import Taylor1, Taylor2, SphericalTaylor2, NonSphericalTaylor2
+
+__all__ = ['Taylor1', 'Taylor2', 'SphericalTaylor2', 'NonSphericalTaylor2']
