+++ sao/convergence_criteria/__init__.py
@@ -1,3 +1,6 @@
 from .fixed import IterationCount, Feasibility
-from .criteria import ObjectiveChange
-from .criteria import VariableChange
+from .change import ObjectiveChange, VariableChange, VariableChangeNorm
+from .kkt import KKT
+
+__all__ = ['IterationCount', 'Feasibility', 'ObjectiveChange', 'VariableChange',
+           'VariableChangeNorm', 'KKT']
