+++ sao/move_limits/adaptive_move_limits.py
@@ -1,5 +1,4 @@
 import numpy as np
-
 from sao.move_limits.move_limit import MoveLimit
