+++ sao/problems/__init__.py
@@ -1,2 +1,4 @@
 from .problem import Problem
 from .subproblem import Subproblem
+
+__all__ = ['Problem', 'Subproblem']
