+++ sao/scaling_strategies/__init__.py
@@ -1,1 +1,4 @@
-from .scaling import Scaling, InitialResponseScaling, InitialObjectiveScaling
+from .scaling import Scaling
+from .initial_scaling import InitialResponseScaling, InitialObjectiveScaling
+
+__all__ = ['InitialResponseScaling', 'InitialObjectiveScaling']
