+++ sao/solvers/__init__.py
@@ -1,6 +1,9 @@
-from sao.solvers.SolverIP_Svanberg import ipsolver
-from sao.solvers.cvxopt_wrapper import cvxopt_solver
-from sao.solvers.method_of_moving_asymptotes import mma
+from sao.solvers.pdip_svanberg import ipsolver
+from sao.solvers.wrappers.cvxopt import cvxopt_solver
+from sao.solvers.wrappers.mma import mma
 from sao.solvers.optimality_criteria import oc, oc1999
-from sao.solvers.primal_dual_interior_point import pdip
-from sao.solvers.scipy_wrapper import scipy_solver
+from sao.solvers.primal_dual_interior_point import pdip, Pdipx, Pdipxy, Pdipxyz
+from sao.solvers.wrappers.scipy import scipy_solver
+
+__all__ = ['ipsolver', 'cvxopt_solver', 'mma', 'oc', 'oc1999',
+           'pdip', 'Pdipx', 'Pdipxy', 'Pdipxyz', 'scipy_solver']
