+++ sao/solvers/pdip_svanberg.py
@@ -1,8 +1,7 @@
-## Imports
 import numpy as np
 
 
-## Svanberg's InteriorPoint solver found in http://www.ingveh.ulg.ac.be/uploads/education/meca-0027-1/MMA_DCAMM_1998.pdf
+# Svanberg's InteriorPoint solver found in http://www.ingveh.ulg.ac.be/uploads/education/meca-0027-1/MMA_DCAMM_1998.pdf
 def ipsolver(problem, x0=None, epsimin=1e-6, max_inner_iter=20, max_lines_iter=20, max_outer_iter=100,
              epsifac=0.9, epsired=0.1, cCoef=1000):
     """
@@ -168,7 +167,7 @@ def ipsolver(problem, x0=None, epsimin=1e-6, max_inner_iter=20, max_lines_iter=2
     return x
 
 
-## Calculates the residual of the relaxed KKT conditions
+# Calculates the residual of the relaxed KKT conditions
 def residual(x, y, z, lam, xsi, eta, mu, zet, s, epsi, a0, a, c, d, problem):
     # Calculating g_j_tilde_value, dg_j_tilde_value and dpsi_dx
     g_j_tilde_value = problem.g(x)
