+++ sao/solvers/wrappers/__init__.py
@@ -1,0 +1,5 @@
+from sao.solvers.wrappers.cvxopt import cvxopt_solver
+from sao.solvers.wrappers.scipy import scipy_solver
+from sao.solvers.wrappers.mma import mma
+
+__all__ = ['cvxopt_solver', 'scipy_solver', 'mma']
