+++ sao/solvers/wrappers/mma.py
@@ -1,5 +1,5 @@
 from sao.approximations import Taylor1
-from sao.convergence_criteria.criteria import VariableChange
+from sao.convergence_criteria.change import VariableChange
 from sao.intervening_variables.mma import MMA02 as MMA
 from sao.move_limits import Bounds, MoveLimit, AdaptiveMoveLimit
 from sao.problems import Subproblem
